# web/tooling/openapi/workforce.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Workforce.Host.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5041/"
    }
  ],
  "paths": {
...
    "/api/v1/recruitment/offers": {
      "get": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferDetailDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferDetailDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfferDetailDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}": {
      "get": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfferTermsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfferTermsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfferTermsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/submit-approval": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/approve": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/issue": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/accept": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/decline": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/offers/{id}/withdraw": {
      "post": {
        "tags": [
          "RecruitmentOffers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConcurrencyActionRequest": {
        "required": [
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "CreateOfferRequest": {
        "required": [
          "applicationId",
          "candidateId",
          "titleEn",
          "titleAr",
          "proposedStartDate",
          "baseSalaryMonthly",
          "currency",
          "allowancesJson",
          "conditionsNote",
          "expiryDate",
          "offerDocumentId"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "titleEn": {
            "type": "string"
          },
          "titleAr": {
            "type": "string"
          },
          "proposedStartDate": {
            "type": "string",
            "format": "date"
          },
          "baseSalaryMonthly": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "allowancesJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "conditionsNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "offerDocumentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "OfferDetailDto": {
        "required": [
          "id",
          "tenantId",
          "legalEntityId",
          "applicationId",
          "candidateId",
          "offerVersionNumber",
          "titleEn",
          "titleAr",
          "proposedStartDate",
          "baseSalaryMonthly",
          "currency",
          "allowancesJson",
          "conditionsNote",
          "status",
          "approvalRequestId",
          "issuedAtUtc",
          "acceptedAtUtc",
          "expiryDate",
          "offerDocumentId",
          "createdAtUtc",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "legalEntityId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "offerVersionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "titleEn": {
            "type": "string"
          },
          "titleAr": {
            "type": "string"
          },
          "proposedStartDate": {
            "type": "string",
            "format": "date"
          },
          "baseSalaryMonthly": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "allowancesJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "conditionsNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "approvalRequestId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "issuedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "acceptedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "expiryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "offerDocumentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
      "OfferStatus": {
        "type": "integer"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}